Charts improvements, external force functions injection, simulation fixes

+++ Statechart.h
@@ -28,7 +28,7 @@ namespace sm
 
 	private:
 		static const int cNumMaxRecords;
-		static const int cOutlineThickness;
+		static const float cOutlineThickness;
 		static const float cMinimumNormDifference;
 
 		sf::FloatRect _viewport;
